Sửa loạt lỗi từ bốn lượt review độc lập, kèm test
Bốn con review chạy song song trên ba tầng, cộng một con thiết kế bộ test
thay vì đi tìm bug. Mọi thứ dưới đây đều đọc được từ code; chỗ nào diễn đạt
được bằng test thì test được viết trước và đỏ trước.

NẶNG NHẤT — playlist không bao giờ tự sang bài. keep-open=yes khiến libmpv
không unload file khi hết (đó chính là thứ giữ khung hình cuối trên màn),
mà MPV_EVENT_END_FILE theo tài liệu chỉ đến SAU khi unload. Nên sự kiện
không bao giờ tới, ReachedEnd không bao giờ bắn, cả nhánh kết-thúc-file là
code chết. Giờ observe property eof-reached, và mọi đường vào trạng thái
Ended đều báo qua một chỗ duy nhất để sự kiện bắn đúng một lần.

Quả mìn ở resolver: RegisterCallingAssembly thêm assembly vào list, đăng ký
resolver cho nó, RỒI mới materialize Lazy vốn duyệt lại chính list đó —
đăng ký lần hai. SetDllImportResolver ném khi trùng, và Lazy cache exception
vĩnh viễn, nên mọi lời gọi libmpv về sau sẽ hỏng chứ không phải lời gọi làm
nó nổ. Chưa nổ chỉ vì cửa sổ tình cờ dựng engine trước renderer. Hai lượt
review độc lập cùng tìm ra chỗ này.

Dispose lần hai thì ném: nó chạm EnterWriteLock trên cái lock mà lần một đã
dispose, trước khi tới được guard _disposed nằm bên trong.

Còn lại: volume/speed không hữu hạn bị từ chối thay vì lưu (Clamp(NaN) là
NaN); trạng thái Ended sống sót qua cú pause ngay sau nó; position âm bị kẹp
khi vào; kéo seek bar không còn bị trả về chỗ cũ (Slider nhả pointer capture
như một phần của cú click bình thường); click vào thanh volume giờ ăn; thoát
fullscreen trả lại cửa sổ maximize; swap chain bám theo đổi scale màn hình;
kiểm mã trả về của ResizeBuffers và Present; resize dựng surface mới trước
khi bỏ cái cũ; marker crash đóng khi có khung hình lên màn chứ không phải
khi constructor return; lỗi do cửa sổ tự báo sống lâu hơn một frame; thanh
điều khiển ở lại khi pipeline video hỏng; dọn log cũ không còn tắt luôn
logging; subscriber log ném không còn giết tiến trình; seek khi chưa có file
bị bỏ qua thay vì ném; tên file log kèm process id; bỏ observe core-idle.

Test: 70 -> 107. Trong đó 5 test viết ra là đỏ ngay, chứng minh bug thật
trước khi sửa. Riêng test "mọi property được observe đều có case trong
reducer" thì đã tồn tại sẵn — nhưng được cấp một danh sách ngoại lệ chứa
đúng cái property đang hỏng. Bỏ danh sách ngoại lệ đi.

Thêm docs/TESTING.md: bốn tầng kiểm thử, cái gì tự động được, cái gì không,
và checklist 12 mục cho người thật.

+++ src/Nocturne.Render/Interop/Egl.cs
@@ -159,8 +159,15 @@ internal static unsafe class Egl
     /// passed to <see cref="GetPlatformDisplay"/> under
     /// <see cref="EGL_PLATFORM_DEVICE_EXT"/>.
     /// </remarks>
+    /// <remarks>
+    /// The attribute list is <c>EGLAttrib*</c> — pointer-sized, unlike the
+    /// <c>EGLint*</c> that <see cref="GetPlatformDisplay"/> takes. Only
+    /// <see langword="null"/> is passed today; declaring it as <c>int*</c>
+    /// would leave the next person to add a real attribute with entries read
+    /// four bytes apart on x64.
+    /// </remarks>
     [DllImport(EglLibrary, EntryPoint = "eglCreateDeviceANGLE")]
-    internal static extern nint CreateDeviceAngle(int deviceType, nint device, int* attributes);
+    internal static extern nint CreateDeviceAngle(int deviceType, nint device, nint* attributes);
 
     [DllImport(EglLibrary, EntryPoint = "eglReleaseDeviceANGLE")]
     internal static extern int ReleaseDeviceAngle(nint device);
